// src/constants/Enums.ts
export enum TABLES {
    ORDER = "order",
    ORDER_PRINTER_COUNTER = "order_printer_counter",
    ORDER_MATERIAL = "order_material",
    MATERIAL = "material",
    CLIENT = "client",
    PRINTER = "printer",
    PRINTER_TYPE = "printer_type",
    PRINTER_MODEL = "printer_model",
    USER = "user",
    COUNTER = "counter",
    TICKET = "ticket",
    TICKET_PRINTER = "ticket_printer",
    UNREGISTERED_CLIENT = "unregistered_client",
    UNREGISTERED_PRINTER = "unregistered_printer",
    PASSWORD_RESET="password_reset",
    ORDER_PRINTER_MATERIAL = "order_printer_material"
}

export enum ORDER_STATUS {
    CLOSED=1,
    OPEN=2
}